chore(suite): separate tx validity timer logic from main review modal component
